Fix the lower bound of the number of keys in the computation of num_buckets
Change the lower bound in the number of keys in the computation of
num_buckets from MAX_CLIENTS * MAX_BUDDIES to 1.

The rationale is that the data file is public, and the number of keys
actually in the data file will be visible to anyone.  If we want to hide
the number of keys in the registration file, we'll need to throw random
keys into the data file as well.  This does not seem important to me at
this time, but it's good to keep in mind.